refactor(vfs): fold session.rs (AgentVfs) into vfs.rs
AgentVfs is small and tightly coupled to the Vfs core, so it doesn't warrant its
own module. Move the `AgentVfs` enum and its impl from session.rs into vfs.rs
(alongside `Vfs`), delete session.rs, and re-export `AgentVfs` from `vfs::vfs`.
No behavior change; the sandbox variant stays feature-gated as before.
