+++ src/vfs/mod.rs
@@ -11,7 +11,6 @@ mod cache;
 pub mod fuse;
 pub mod path;
 pub mod resource;
-pub mod session;
 
 /// Sandbox-only frontend (host forward server + in-guest forwarder).
 #[cfg(feature = "sandbox")]
@@ -24,10 +23,8 @@ pub use accessor::{GDriveConfig, NotionConfig, S3Config};
 pub use fuse::VfsMount;
 pub use path::VPath;
 pub use resource::{DirEntry, FileKind, FileStat, Resource, S3Resource};
-pub use session::AgentVfs;
-pub use vfs::{Mount, MountSpec, ProviderConfig, Vfs, VfsConfig};
-
 // Keep the public paths `crate::vfs::{VfsForward, bootstrap_guest_forwarder}`
 // stable; they live in the sandbox submodule but are only present with the feature.
 #[cfg(feature = "sandbox")]
 pub use sandbox::{VfsForward, bootstrap_guest_forwarder};
+pub use vfs::{AgentVfs, Mount, MountSpec, ProviderConfig, Vfs, VfsConfig};
